fix and reorganize interrupts

+++ src/io.h
@@ -4,7 +4,7 @@
 #include "const.h"
 
 u8 inb(u16 port);
-void outb(u16 port, u16 data);
+void outb(u16 port, u8 data);
 void io_wait();
 
 #endif
